pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "cmake>=3.18", "pybind11>=2.11"]
build-backend = "setuptools.build_meta"

[project]
name = "chisel"
version = "0.1.0"
description = "Multi-view 3D Reconstruction Pipeline with ETH3D evaluation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "opencv-python>=4.5",
    "torch>=2.0",
    "scipy>=1.7",
    "pyyaml>=6.0",
    "matplotlib>=3.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black", "mypy", "ruff"]
viz = ["open3d>=0.17", "trimesh"]

[tool.setuptools.packages.find]
where = ["python"]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
pythonpath = ["python"]

[tool.ruff]
line-length = 100
target-version = "py39"
